[server] Move the logger module form netlib to util
