+++ src/server/netlib/io_event.cc
@@ -9,7 +9,7 @@
 
 #include "io_event.h"
 
-#include "log.h"
+#include "../util/log.h"
 
 using namespace netlib;
