+++ src/server/netlib/socket.cc
@@ -12,7 +12,7 @@
 // TODO chosed by os
 #include <fcntl.h>
 
-#include "log.h"
+#include "../util/log.h"
 
 using namespace netlib;
