+++ src/server/netlib/tcp_server.h
@@ -10,10 +10,10 @@
 #ifndef MRPROXYER_SRC_SERVER_NETLIB_TCPSERVER_H_
 #define MRPROXYER_SRC_SERVER_NETLIB_TCPSERVER_H_
 
+#include "../util/log.h"
 #include "../util/non_copyable_movable.h"
 #include "acceptor.h"
 #include "io_watcher.h"
-#include "log.h"
 #include "netlib.h"
 #include "tcp_event.h"
